checker Cyrillic comments
